requirements.txt
torch>=2.0.0
transformers>=4.30.0
Pillow>=9.0.0
matplotlib>=3.5.0
numpy>=1.21.0
hydra-core>=1.3.0
omegaconf>=2.3.0
